refactor(front): splitted repository manager & create form
